+++ dart/geobase/lib/geobase.dart
@@ -91,6 +91,7 @@ export 'src/meta/time/interval.dart';
 export 'src/meta/time/temporal.dart';
 
 // projections
+export 'src/projections/ellipsoidal/ellipsoidal_projection_adapter.dart';
 export 'src/projections/wgs84/wgs84.dart';
 
 // tiling